requirements.txt
clip==0.2.0
clip_anytorch==2.4.0
Coop==3.0.0
h5py==3.7.0
numpy==1.23.0
pandas==1.4.2
pyinstaller==5.13.0
Pillow==9.5.0
PyQt6==6.4.0
PyQt6_sip==13.4.0
scikit_learn==1.1.1
torch==1.12.0
tqdm==4.64.0
transformers==4.23.1
XlsxWriter==3.0.3
